+++ src/marketplace/Offers.Library/Service/OfferSetupService.cs
@@ -217,20 +217,26 @@ public class OfferSetupService : IOfferSetupService
             offerDetails.CompanyName,
             offerDetails.OfferName
         });
-        await _notificationService.CreateNotifications(
+        
+        Guid? creatorId = offerDetails.CompanyUserId != Guid.Empty ? offerDetails.CompanyUserId : null;
+        var userIdsOfNotifications = await _notificationService.CreateNotifications(
             itAdminRoles,
-            offerDetails.CompanyUserId != Guid.Empty ? offerDetails.CompanyUserId : null,
+            creatorId,
             new (string?, NotificationTypeId)[]
             {
-                (null, NotificationTypeId.TECHNICAL_USER_CREATION),
+                (JsonSerializer.Serialize(new { offerDetails.OfferId, offerDetails.OfferName }), NotificationTypeId.TECHNICAL_USER_CREATION),
                 (notificationContent, appSubscriptionActivation)
             },
-            offerDetails.CompanyId).ConfigureAwait(false);
+            offerDetails.CompanyId).ToListAsync().ConfigureAwait(false);
 
-        _portalRepositories.GetInstance<INotificationRepository>().CreateNotification(offerDetails.RequesterId, appSubscriptionActivation, false, notification =>
+        if (!userIdsOfNotifications.Contains(offerDetails.RequesterId))
         {
-            notification.Content = notificationContent;
-        });
+            _portalRepositories.GetInstance<INotificationRepository>().CreateNotification(offerDetails.RequesterId, appSubscriptionActivation, false, notification =>
+            {
+                notification.Content = notificationContent;
+                notification.CreatorUserId = creatorId;
+            });
+        }
     }
 
     private async Task SendMail(string basePortalAddress, string userName, string requesterEmail, string? offerName)
@@ -245,4 +251,6 @@ public class OfferSetupService : IOfferSetupService
             .SendMails(requesterEmail, mailParams, new List<string> {"subscription-activation"})
             .ConfigureAwait(false);
     }
+
+    internal record CreateTechnicalUserData(Guid CompanyId, string? OfferName, string? Bpn, string TechnicalUserName, bool EnhanceTechnicalUserName);
 }