+++ CHANGELOG.md
@@ -2,6 +2,50 @@
 
 New features, fixed bugs, known defects and other noteworthy changes to each release of the Catena-X Portal Backend.
 
+## 1.3.0-RC4
+
+### Change
+* Seeding Data: update agreement name for app marketplace offers - CX Conformity
+
+### Feature
+* Services Service: released document delete endpoint DELETE: /api/services/servicerelease/documents/{documentId}
+
+### Technical Support
+n/a
+
+### Bugfix
+n/a
+
+## 1.3.0-RC3
+
+### Change
+n/a
+
+### Feature
+n/a
+
+### Technical Support
+* updated sonarcloud workflow: use repo variables for project key and organization
+
+### Bugfix
+* Notification Service: the notifications will only get created once per request for each user and notification type
+
+## 1.3.0-RC2
+
+### Change
+* Services Service
+  * POST: /api/services/addservice - added "providerUri" inside the request body to store the provider url
+  * PUT: /api/services/{serviceId} - added "providerUri" for endpoint to store the provider url
+
+### Feature
+n/a
+
+### Technical Support
+n/a
+
+### Bugfix
+n/a
+
 ## 1.3.0-RC1
 
 ### Change